Feb. 13
This is the update where i have the reviewhelpervc basically done. each modally presented vc sends the info from itself back to that vc and the 4 or 5 review parts compile in this vc. The next step is to have all the review parts get stored in the database. Commit at that point, then go to the feed vc and put in the code to retrieve the change in the database and copy that to a cell.