Mar. 4
This is the upload before I try and combine the two types of reviews into one tableview. I have the code for the reviewpost in the cellForRowAt function, but I havent done anything for the reviewpost in the viewdidload/observe method. Im gonna try and get it so that I can get the reviewposts up into the tableview without the reviews and then try to use that to combine the two so right now im gonna silence the observe method that contains review and make a new one for reviewposts.