July 6
This is the upload after completing the timesincepost stuff. (Date vs Date() was huge. I should look into more of that syntax stuff soon. I am going to work on finding/saving coffee shops next. This will be a big part in the first step of involving more of a social media aspect into the app.