Mar 19
The upload before I try and change the genericposttvc to be abelt o have likes and comments. This could be a lot so im saving it in case  trying to use the followingArray on every tableviewcell might affect it. Even though it is already being used and the only porblem is that it doesnt change automatically which is something else I have to look into.