Mar. 24
This is the commit before I finished the new User liked posts path. the porblem is that the code doesnt recognize if a post has already been liked by the user so it just keeps adding the like number. I am saving this one because what I did on lines 203-215 in shop review cell plus a few other places was kind of crafty and i dont want to lose it in case the problem comes up again. Using datasnapshot and dictionary then getting the value of the dictionary to put into an array is the code i am saving